reorganize package into directory per format
