+++ packages/era/src/era1/era1.ts
@@ -7,9 +7,8 @@ import {
 } from '@ethereumjs/util'
 import * as ssz from 'micro-eth-signer/ssz'
 
+import { EpochAccumulator, Era1Types, VERSION, formatEntry, readEntry } from '../index.ts'
 import { blockFromTuple, parseBlockTuple, readBlockTupleAtOffset } from './blockTuple.ts'
-import { formatEntry, readEntry } from './e2store.ts'
-import { EpochAccumulator, Era1Types, VERSION } from './types.ts'
 
 /**
  * Format era1 from epoch of history data
