+++ packages/era/src/index.ts
@@ -1,8 +1,9 @@
 import { readFileSync } from 'fs'
 
-export * from './blockTuple.ts'
+export * from './era/index.ts'
 export * from './e2store.ts'
-export * from './era1.ts'
+export * from './era1/index.ts'
+export * from './e2hs/index.ts'
 export * from './exportHistory.ts'
 export * from './snappy.ts'
 export * from './types.ts'
